+++ node/utils/index.ts
@@ -1,5 +1,11 @@
 import { parseOrderFormToProtocol } from './parseOrderFormToProtocol'
 import { getOrderFormId } from './getOrderFormId'
 import { applyApportionment } from './applyApportionment'
+import { validateIndexConsistency } from './validateIndexConsistency'
 
-export { parseOrderFormToProtocol, getOrderFormId, applyApportionment }
+export {
+  parseOrderFormToProtocol,
+  getOrderFormId,
+  applyApportionment,
+  validateIndexConsistency,
+}